+++ CHANGES.rst
@@ -13,10 +13,13 @@ Breaking changes:
 - Moved code around and deprecated old locations in ``Products/CMFPlone/patterns/__init__``.
   This goes together with same pattern settings changes in ``plone.app.layout.globals.pattern_settings``.
   Also moved general usable ``./patterns/utils/get_portal`` to ``./utils/.get_portal``.
-  Depreacted ``./patterns/utils/get_portal`` and ``./patterns/utils/get_portal``.
+  Deprecated ``./patterns/utils/get_portal`` and ``./patterns/utils/get_portal``.
   [jensens]
 
 New features:
+
+- Added setting to editing controlpanel to enable limit of keywords to the current navigation root.
+  [jensens]
 
 - Make login modal dialog follow any redirects set while processing the login request.
   [fulv]